Move inference accessible label composition into inferenceStatus.
Fleet torp module should own fleet semantics only; badge label assembly now lives next to inferenceAccessibleLabel.
